Move another helper to the pile
